# Clustered lighting optimisations - removed WebGL1 compatibility solutions