… examples/src/examples/graphics/clustered-spot-shadows.controls.mjs
… examples/src/examples/graphics/clustered-spot-shadows.example.mjs
… src/scene/renderer/renderer.js
… src/scene/shader-lib/chunks/chunks.js
… src/scene/shader-lib/chunks/lit/frag/clusteredLight.js
… src/scene/shader-lib/chunks/lit/frag/clusteredLightCookies.js
… src/scene/shader-lib/chunks/lit/frag/float-unpacking.js
… src/scene/shader-lib/chunks/lit/frag/lighting/lighting.js