## Main goals done
- lights were stored in an 8bit and a float texture. 8bit texture is now removed and so one additional texture slot is available
- WebGL1 didn't have bit packing / uint arithmetic, and so many 1bit data were stored in 8bits of 8bit texture. Those are now bit packed, requiring less bandwidth used by the texture sampling
- using hardware half->float decompression to store data, instead of manual decompression
- storing data pre-cached in half float formats in a light class instead of compressing each frame

I expect slight CPU performance win, but this was not tested.

## GPU improvements (4K on M1 Mac):

11% improvement in this example
before: 1.8ms
now: 1.6ms
![Screenshot 2025-03-20 at 16 58 57](https://github.com/user-attachments/assets/40258ec8-84c3-470e-967e-9c579f500af3)

16% improvement in this example
before: 4.2ms
now: 3.5ms
![Screenshot 2025-03-20 at 16 59 52](https://github.com/user-attachments/assets/ceb85295-2bb2-487a-b9cd-0e9cf32ddce6)